+++ src/client/java/com/chedidandrew/smartresourcedrops/client/SmartResourceDropsModMenuIntegration.java
@@ -7,7 +7,7 @@ import com.terraformersmc.modmenu.api.ModMenuApi;
  * Optional Mod Menu integration.
  *
  * <p>This class is only discovered by Mod Menu through the custom "modmenu"
- * entrypoint. Resource Multiplier does not require Mod Menu to run.</p>
+ * entrypoint. Smart Resource Multiplier does not require Mod Menu to run.</p>
  */
 public final class SmartResourceDropsModMenuIntegration implements ModMenuApi {
     @Override